+++ src/main/java/io/supertokens/webserver/api/HandshakeAPI.java
@@ -75,6 +75,11 @@ public class HandshakeAPI extends WebserverAPI {
             result.addProperty("refreshTokenPath", Config.getConfig(main).getRefreshAPIPath());
             result.addProperty("enableAntiCsrf", Config.getConfig(main).getEnableAntiCSRF());
             result.addProperty("accessTokenBlacklistingEnabled", Config.getConfig(main).getAccessTokenBlacklisting());
+            if (!this.version.equals("1.0")) {
+                result.addProperty("cookieSameSite", Config.getConfig(main).getCookieSameSite());
+                result.addProperty("idRefreshTokenPath", Config.getConfig(main).getAccessTokenPath());
+                result.addProperty("sessionExpiredStatusCode", Config.getConfig(main).getSessionExpiredStatusCode());
+            }
             super.sendJsonResponse(200, result, resp);
         } catch (StorageQueryException | StorageTransactionLogicException e) {
             throw new ServletException(e);