+++ src/main/java/io/supertokens/webserver/api/JWTDataAPI.java
@@ -34,16 +34,12 @@
 
 package io.supertokens.webserver.api;
 
-import com.google.gson.Gson;
+import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 import io.supertokens.Main;
-import io.supertokens.exceptions.TokenTheftDetectedException;
 import io.supertokens.exceptions.UnauthorisedException;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
-import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicException;
 import io.supertokens.session.Session;
-import io.supertokens.session.info.SessionInformationHolder;
 import io.supertokens.webserver.InputParser;
 import io.supertokens.webserver.WebserverAPI;
 
@@ -52,65 +48,76 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class RefreshSessionAPI extends WebserverAPI {
-    private static final long serialVersionUID = 7142317017402226537L;
+public class JWTDataAPI extends WebserverAPI {
+    private static final long serialVersionUID = -4989144736402314280L;
 
-    public RefreshSessionAPI(Main main) {
+    public JWTDataAPI(Main main) {
         super(main);
     }
 
     @Override
     public String getPath() {
-        return "/session/refresh";
+        return "/jwt/data";
     }
 
     @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
-        // Do not modify after and including this line
-        if (sendRandomUnauthorisedIfDevLicenseAndSomeTimeHasPassed(resp)) {
-            return;
+    protected void doPut(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
+        if (this.version.equals("1.0")) {
+            throw new ServletException(
+                    new BadRequestException("CDI version not supported"));
         }
-        // Do not modify before and including this line
         JsonObject input = InputParser.parseJsonObjectOrThrowError(req);
-        String refreshToken = InputParser.parseStringOrThrowError(input, "refreshToken", false);
-        assert refreshToken != null;
+
+        String sessionHandle = InputParser.parseStringOrThrowError(input, "sessionHandle", false);
+        assert sessionHandle != null;
+
+        JsonObject userDataInJWT = InputParser.parseJsonObjectOrThrowError(input, "userDataInJWT", false);
+        assert userDataInJWT != null;
 
         try {
-            SessionInformationHolder sessionInfo = Session
-                    .refreshSession(main, refreshToken);
+            Session.updateSession(main, sessionHandle, null, userDataInJWT, null);
 
-            JsonObject result = new JsonParser().parse(new Gson().toJson(sessionInfo)).getAsJsonObject();
-
-            if (this.version.equals("1.0")) {
-                result.getAsJsonObject("accessToken").remove("sameSite");
-                result.getAsJsonObject("refreshToken").remove("sameSite");
-                result.getAsJsonObject("idRefreshToken").remove("sameSite");
-                result.getAsJsonObject("idRefreshToken").remove("cookiePath");
-                result.getAsJsonObject("idRefreshToken").remove("cookieSecure");
-                result.getAsJsonObject("idRefreshToken").remove("domain");
-            }
+            JsonObject result = new JsonObject();
 
             result.addProperty("status", "OK");
             super.sendJsonResponse(200, result, resp);
-        } catch (StorageQueryException | StorageTransactionLogicException e) {
+
+        } catch (StorageQueryException e) {
             throw new ServletException(e);
         } catch (UnauthorisedException e) {
             JsonObject reply = new JsonObject();
             reply.addProperty("status", "UNAUTHORISED");
             reply.addProperty("message", e.getMessage());
             super.sendJsonResponse(200, reply, resp);
-        } catch (TokenTheftDetectedException e) {
+        }
+    }
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
+        if (this.version.equals("1.0")) {
+            throw new ServletException(
+                    new BadRequestException("CDI version not supported"));
+        }
+        String sessionHandle = InputParser.getQueryParamOrThrowError(req, "sessionHandle", false);
+        assert sessionHandle != null;
+
+        try {
+            JsonElement jwtPayload = Session.getJWTData(main, sessionHandle);
+
+            JsonObject result = new JsonObject();
+
+            result.addProperty("status", "OK");
+            result.add("userDataInJWT", jwtPayload);
+            super.sendJsonResponse(200, result, resp);
+
+        } catch (StorageQueryException e) {
+            throw new ServletException(e);
+        } catch (UnauthorisedException e) {
             JsonObject reply = new JsonObject();
-            reply.addProperty("status", "TOKEN_THEFT_DETECTED");
-
-            JsonObject session = new JsonObject();
-            session.addProperty("handle", e.sessionHandle);
-            session.addProperty("userId", e.userId);
-            reply.add("session", session);
-
+            reply.addProperty("status", "UNAUTHORISED");
+            reply.addProperty("message", e.getMessage());
             super.sendJsonResponse(200, reply, resp);
         }
-
-        super.saveDeviceDriverInfoIfPresent(input);
     }
+
 }