+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2020-04-30
+### Added
+- Compatibility with CDI 2.0
+- API versions
+- SameSite cookie option
+- Updating of JWT payload
+- Session expired status code configuration
+- Partial lmrt support
+
 ## [2.0.0] - 2020-04-07
 ### Added
 - Compatibility with NoSQL databases like MongoDB